Bug fix if the final snapshot was not at the end of the simulation
The code previously had a problem if the ending snapshot was no at the end of
the simulation, this has bee fixed. In addition added in a catch so if there
are no halos then it does not try to output the final tree, that would also
contain no halos.