Enabled WhereWolf to restart by creating and reading restart file
At the end of each run WhereWolf will create a restart file for each MPI thread, this is then automatically checked for upon restart and read in if it is found so it can continue as if there was no break in the run. There is a slight change to the ui now where -n specifies the number of snapshots and -s specifies the snapshot offset if there is one, this means that this does not reqiure the input filelists to change.