Small bug fix
Because of the if(isnap<fsnap) the final StartTrack would not run and so the restart file would not contain the data for the halos that started to be tracked in the previous snapshot. This has now been fixed where the tree is now used to define when at the final snapshot is, when there is no descendant dataset in the file.