Bug fix to IDs and an update to the stop tracking criterion
I was adjusting the ID for the VELOCIraptor halos when they were a descendant of a WhereWolf halo which would mean it would point to the wrong halo, this has been corrected for. In addition WhereWolf will now stop tracking when the number of most bound particles drops below 5 since a meaningful center for a halo cannot be found.