Update so more VELOCIraptor halo are searched when merging and bug fix
WhereWolf now searches more VELOCIraptor halos looking for any connection that it merged with another. In addition there is a bug fix where if the number of halo searched is greater than the current number of halos the kdtree would return the number of halos in that snapshot for indexes great than the number of halos, giving a index error.