Code cleanup and WhereWolf searches for VELOCIraptor halo to connect to
Another script has been created to cleanup the code a bit. In addtion WhereWolf is now able to find a matching VELOCIraptor halo and sees if it already has a connection. If not then the WhereWolf halo is connected if above some noise limit. Otherwise if a connection already exists and if the connection to the WhereWolf halo is better by 20% then the connection is switched.