Update to the output
WhereWolf does not output to the VELOIraptor catalog but now will output to a seperate directory, so does not update the VELOCIraptor catalog to include the WhereWolf. This means to load the WhereWolf halos a additional function will need to be called to read and append it to the halodata. This function is under development and will be added soon.